requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
appdirs=1.4.4=pyh9f0ad1d_0
argon2-cffi=20.1.0=py38h2bbff1b_1
asn1crypto=1.4.0=pyh9f0ad1d_0
async_generator=1.10=pyhd3eb1b0_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
blosc=1.21.0=h19a0ad4_0
brotlipy=0.7.0=py38h294d835_1001
bzip2=1.0.8=he774522_0
ca-certificates=2021.1.19=haa95532_1
certifi=2020.12.5=py38haa95532_0
cffi=1.14.5=py38hd8c33c5_0
chardet=4.0.0=py38haa244fe_1
click=7.1.2=pyh9f0ad1d_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyhd3eb1b0_0
croniter=0.3.36=pyhd8ed1ab_0
cryptography=3.4.4=py38hb7941b4_0
cudatoolkit=11.1.1=heb2d755_7
cycler=0.10.0=py38_0
cytoolz=0.11.0=py38h294d835_3
dask-core=2021.2.0=pyhd8ed1ab_0
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
distributed=2021.2.0=py38haa244fe_0
docker-py=4.4.3=py38haa244fe_0
docker-pycreds=0.4.0=py_0
entrypoints=0.3=py38_0
freetype=2.10.4=hd328e21_0
hdf5=1.10.4=h7ebc959_0
heapdict=1.0.1=py_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyh9f0ad1d_0
importlib-metadata=3.7.3=py38haa95532_1
importlib_metadata=3.7.3=hd3eb1b0_1
importlib_resources=5.1.0=py38haa244fe_0
intel-openmp=2020.3=h57928b3_311
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.21.0=py38hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jedi=0.17.0=py38_0
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py38_7
jupyter_client=6.1.7=py_0
jupyter_console=6.3.0=pyhd3eb1b0_0
jupyter_core=4.7.1=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=py38hd77b12b_0
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
liblapack=3.9.0=8_mkl
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.2.0=he0120a3_0
libuv=1.40.0=he774522_0
lz4-c=1.9.3=h2bbff1b_0
lzo=2.10=he774522_2
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py38he774522_0
marshmallow=3.10.0=pyhd8ed1ab_0
marshmallow-oneofschema=2.0.1=py_0
matplotlib=3.3.4=py38haa95532_0
matplotlib-base=3.3.4=py38h49ac443_0
mistune=0.8.4=py38he774522_1000
mkl=2020.4=hb70f87d_311
mkl-service=2.3.0=py38h196d8e1_0
mock=4.0.3=pyhd3eb1b0_0
msgpack-python=1.0.2=py38hbd9d945_1
msys2-conda-epoch=20160418=1
mypy_extensions=0.4.3=py38haa244fe_3
natsort=7.1.1=pyhd8ed1ab_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py38_0
nbformat=5.1.2=pyhd3eb1b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
ninja=1.10.2=py38h6d14046_0
notebook=6.2.0=py38haa95532_0
numexpr=2.7.3=py38hcbcaa1e_0
numpy=1.20.1=py38h0cc643e_0
olefile=0.46=py_0
openssl=1.1.1j=h2bbff1b_0
packaging=20.9=pyh44b312d_0
pandas=1.2.2=py38h4c96930_0
pandoc=2.12=haa95532_0
pandocfilters=1.4.3=py38haa95532_1
parso=0.8.1=pyhd3eb1b0_0
pendulum=2.1.2=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.2=py38h4fa10fc_0
pip=21.0.1=py38haa95532_0
prefect=0.14.9=pyhd8ed1ab_0
prometheus_client=0.9.0=pyhd3eb1b0_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
psutil=5.8.0=py38h294d835_1
pycparser=2.20=pyh9f0ad1d_2
pygments=2.8.1=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py38ha925a31_4
pyrsistent=0.17.3=py38he774522_0
pysocks=1.7.1=py38haa244fe_3
pytables=3.6.1=py38ha5be198_0
python=3.8.5=h5fd99cc_1
python-box=5.3.0=pyhd8ed1ab_0
python-dateutil=2.8.1=py_0
python-slugify=4.0.1=pyh9f0ad1d_0
python_abi=3.8=1_cp38
pytorch=1.8.0=py3.8_cuda11.1_cudnn8_0
pytz=2021.1=pyhd8ed1ab_0
pytzdata=2020.1=pyh9f0ad1d_0
pywin32=227=py38h294d835_1
pywinpty=0.5.7=py38_0
pyyaml=5.4.1=py38h294d835_0
pyzmq=20.0.0=py38hd77b12b_1
qt=5.9.7=vc14h73c81de_0
qtconsole=5.0.2=pyhd3eb1b0_0
qtpy=1.9.0=py_0
requests=2.25.1=pyhd3deb0d_0
ruamel.yaml=0.16.12=py38h294d835_2
ruamel.yaml.clib=0.2.2=py38h294d835_2
scikit-learn=0.24.1=py38ha09990b_0
scipy=1.6.1=py38h14eb087_0
seaborn=0.11.1=pyhd3eb1b0_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py38haa95532_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=pyh9f0ad1d_0
sortedcontainers=2.3.0=pyhd8ed1ab_0
sqlite=3.33.0=h2a8f88b_0
tabulate=0.8.8=pyhd8ed1ab_0
tblib=1.6.0=py_0
terminado=0.9.3=py38haa95532_0
testpath=0.4.4=pyhd3eb1b0_0
text-unidecode=1.3=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
toml=0.10.2=pyhd8ed1ab_0
toolz=0.11.1=py_0
torchaudio=0.8.0=py38
torchvision=0.9.0=py38_cu111
tornado=6.1=py38h294d835_1
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
unidecode=1.2.0=pyhd8ed1ab_0
urllib3=1.26.3=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
websocket-client=0.57.0=py38haa244fe_4
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38_0
win_inet_pton=1.1.0=py38haa244fe_2
wincertstore=0.2=py38_0
winpty=0.4.3=4
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zeromq=4.3.3=ha925a31_3
zict=2.0.0=py_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
